[FE feat] #207 친구 스토리 보기 기능 추가-컴포넌트 분리 후 기능 적용
